requirements.txt
Flask==2.2.2
#git+https://github.com/Merubokkusu/Discord-S.C.U.M.git#egg=discum
pillow
